wpi_uitkeringsfraude/Repo wpi-onderzoekswaardigheid-aanvraag - based on dev branch commit 069435abe/pyproject.toml
[tool.poetry]
name = "wpi_onderzoekswaardigheid_aanvraag"
# Note that the version number here must always be related to the version of the input dataset that is archived in the AzureML storage account.
# Models trained on input dataset 1.0.0 should have version numbers in the form of 1.x.x.
# Models trained on input dataset 2.0.0 should have version numbers in the form of 2.x.x. And so on.
version = "1.0.0"
description = ""
authors = []

[[tool.poetry.source]]
name = "gemeente"
url = "https://pkgs.dev.azure.com/CloudCompetenceCenter/Datateam-Sociaal/_packaging/team-AA/pypi/simple/"
secondary = true

[tool.poetry.dependencies]
python = "~3.8"
fraude_preventie = {version = "1.10.12", source = "gemeente"}
category-encoders = {version = "^2.4.0", source = "PyPI"}
interpret = {version = "^0.2.7", source = "PyPI"}
interpret-core = {version = "0.2.7", source = "PyPI"}
pandas = {version = "1.4.4", source = "PyPI"}
pre-commit = {version = "2.21.0", source = "PyPI"}
psycopg2 = {version = "^2.8.6", source = "PyPI"}
scikit-learn = {version = "^0.24", source = "PyPI"}
sklearn-pandas = {version = "^2.0.3", source = "PyPI"}
tables = {version = "^3.6.1", source = "PyPI"}
tabulate = {version = "^0.8.9", source = "PyPI"}
tensorflow = {version = "^2.9.3", source = "PyPI"}
keras = {version = "^2.8.0", source = "PyPI"}
xgboost = {version = "^1.5.2", source = "PyPI"}
Jinja2 = {version = "3.0.3", source = "PyPI"}
fastapi = {version = "^0.92.0", source = "PyPI"}
uvicorn = {version = "^0.20.0", source = "PyPI"}
opencensus-ext-azure = {version = "^1.1.7", source = "PyPI"}
azure-keyvault = {version = "^4.2.0", source = "PyPI"}
azureml-core = {version = "1.49.0", source = "PyPI"}
ez_setup = {version = "^0.9", source = "PyPI"}
wheel = {version = "^0.38.1", source = "PyPI"}
joblib = {version = "^1.2.0", source = "PyPI"}
protobuf = {version = "3.18.3", source = "PyPI"}
Pillow = {version = "9.3.0", source = "PyPI"}
dash = {version = "^2.3.1", source = "PyPI"}
cryptography = {version = "39.0.1", source = "PyPI"}
orjson = {version = "^3.8.5", source = "PyPI"}
setuptools = {version = "<60.0.0", source = "PyPI"}

[tool.poetry.dev-dependencies]
amla-toolkit = {version = "0.5.1", source = "gemeente"}
bias_collection = {version = "0.1.7", source = "gemeente"}
great-expectations = {version = "^0.13.25", source = "PyPI"}
matplotlib = {version = "^3.6.3", source = "PyPI"}
missingno = {version = "0.5.0", source = "PyPI"}
nbstripout = {version = "^0.3.9", source = "PyPI"}
nest-asyncio = {version = "^1.4.3", source = "PyPI"}
pdoc3 = {version = "^0.10.0", source = "PyPI"}
plotly = {version = "^5.0.0", source = "PyPI"}
pytest = {version = "^7.0.1", source = "PyPI" }
pytest-azurepipelines = {version = "1.0.3", source = "PyPI" }
pytest-cov = {version = "^4.0.0", source = "PyPI" }
pytest-mock = {version = "^3.6.1", source = "PyPI"}
pretty-errors = {version = "^1.2.24", source = "PyPI"}
nbconvert = {version = "^6.4.4", source = "PyPI"}
openpyxl = {version = "^3.0.9", source = "PyPI"}
SQLAlchemy = {version = "^1.4.29", source = "PyPI"}

[tool.isort]
profile = "black"
multi_line_output = 3

[build-system]
requires = ["poetry-core>=1.0"]
build-backend = "poetry.core.masonry.api"

